Get rid of "Url" as defined by browsers, replace with "Segments"
Implement URL splitting in Elm so that we are guaranteed to get the
same behavior whether the URL is parsed on the client or server.

This required rather extensive changes to the docs. I think the docs
are a bit less clear, but I’m hoping the overall balance is nicer.